Add files via upload
This data set is created from public data on the Government of British Columbia's website on agreements with First Nations as of August 25, 2019.  There are 2 very similar columns:  Community and Name.  The Name column is a duplicate and transformation of Community to help clean it up a bit.  Eventually, I would like to assign locations to be able to map the location of the agreements.